image-classifier-api/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
torch==2.3.0
torchvision==0.18.0
pillow==10.3.0
pydantic==2.7.1
python-multipart==0.0.9
numpy==1.26.4